# Repair 404 performance regression by using a shared TopicInformer