--- a/http-gateway/pkg/server/messages_handler.go
+++ b/http-gateway/pkg/server/messages_handler.go
@@ -41,12 +41,7 @@ func (g *gateway) messagesHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	topicExists, err := g.topicExistenceChecker.TopicExists(defaultNamespace, topicName)
-	if err != nil {
-		errMsg := fmt.Sprintf("while checking to see if there was a Riff topic '%s', an unexpected error occurred: %+v", topicName, err)
-		http.Error(w, errMsg, http.StatusInternalServerError)
-		return
-	}
+	topicExists:= g.topicExistenceChecker.TopicExists(defaultNamespace, topicName)
 	if !topicExists {
 		errMsg := fmt.Sprintf("could not find Riff topic '%s'", topicName)
 		http.Error(w, errMsg, http.StatusNotFound)