--- a/http-gateway/pkg/server/topic_existence_checker.go
+++ b/http-gateway/pkg/server/topic_existence_checker.go
@@ -57,7 +57,10 @@ func NewRiffTopicExistenceChecker(clientSet *versioned.Clientset) TopicExistence
 
 			key, err := cache.MetaNamespaceKeyFunc(obj)
 			if err != nil {
+				// It is likely that the key is faulty, but we cannot signal an error.
+				// To prevent errors for processing a bad key, we return after logging.
 				log.Printf("AddFunc had an error for key '%s': %+v", err)
+				return
 			}
 
 			knownTopics[key] = ignoredValue{}
@@ -70,7 +73,10 @@ func NewRiffTopicExistenceChecker(clientSet *versioned.Clientset) TopicExistence
 
 			key, err := cache.DeletionHandlingMetaNamespaceKeyFunc(obj)
 			if err != nil {
+				// It is likely that the key is faulty, but we cannot signal an error.
+				// To prevent errors for processing a bad key, we return after logging.
 				log.Printf("DeleteFunc had an error for key '%s': %+v", err)
+				return
 			}
 
 			delete(knownTopics, key)