--- a/http-gateway/pkg/server/topic_existence_checker.go
+++ b/http-gateway/pkg/server/topic_existence_checker.go
@@ -55,6 +55,7 @@ func NewRiffTopicExistenceChecker(clientSet *versioned.Clientset) TopicExistence
 			mutex.Lock()
 			defer mutex.Unlock()
 
+			//TODO: implement riff-specific KeyFunc https://github.com/projectriff/riff/pull/558#discussion_r184437224
 			key, err := cache.MetaNamespaceKeyFunc(obj)
 			if err != nil {
 				// It is likely that the key is faulty, but we cannot signal an error.
@@ -71,6 +72,7 @@ func NewRiffTopicExistenceChecker(clientSet *versioned.Clientset) TopicExistence
 			mutex.Lock()
 			defer mutex.Unlock()
 
+			//TODO: implement riff-specific KeyFunc https://github.com/projectriff/riff/pull/558#discussion_r184437224
 			key, err := cache.DeletionHandlingMetaNamespaceKeyFunc(obj)
 			if err != nil {
 				// It is likely that the key is faulty, but we cannot signal an error.