## Background

The issue reported at projectriff/riff#553 showed that topic checking introduced a performance regression of 2 orders of magnitude. This renders the s1p sample unusable. The commits in this PR change the implementation to use a callback-with-cache mechanism provided by Kubernetes APIs (shared informers).

To validate performance improvement, I compared the `alwaysTrueTopicExistenceChecker` and `RiffTopicExistenceChecker` using `ab2`. Not very scientific, but a useful sanity check. My test case shows the two checking strategies now perform approximately identically, suggesting they are hitting some bottleneck unrelated to topic lookup.

## alwaysTrueTopicExistenceChecker

```
ab -H "Content-Type: text/plain" -n 100000 -c 100 -p in.txt http://192.168.39.8:30788/requests/numbers
...
Concurrency Level:      100
Time taken for tests:   19.363 seconds
Complete requests:      100000
Failed requests:        0
Total transferred:      10400000 bytes
Total body sent:        17700000
HTML transferred:       300000 bytes
Requests per second:    5164.55 [#/sec] (mean)
Time per request:       19.363 [ms] (mean)
Time per request:       0.194 [ms] (mean, across all concurrent requests)
Transfer rate:          524.53 [Kbytes/sec] received
                        892.70 kb/s sent
                        1417.23 kb/s total

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.2      0       5
Processing:     5   19  38.6     17    1268
Waiting:        5   19  38.6     17    1268
Total:          5   19  38.7     18    1270

Percentage of the requests served within a certain time (ms)
  50%     18
  66%     19
  75%     20
  80%     21
  90%     24
  95%     26
  98%     29
  99%     32
 100%   1270 (longest request)
```

## RiffTopicExistenceChecker using a TopicInformer

```
ab -H "Content-Type: text/plain" -n 100000 -c 100 -p in.txt http://192.168.39.8:32408/requests/numbers
...
Concurrency Level:      100
Time taken for tests:   19.131 seconds
Complete requests:      100000
Failed requests:        0
Total transferred:      10400000 bytes
Total body sent:        17700000
HTML transferred:       300000 bytes
Requests per second:    5227.05 [#/sec] (mean)
Time per request:       19.131 [ms] (mean)
Time per request:       0.191 [ms] (mean, across all concurrent requests)
Transfer rate:          530.87 [Kbytes/sec] received
                        903.50 kb/s sent
                        1434.38 kb/s total

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.2      0       5
Processing:     5   19  38.1     17    1247
Waiting:        5   19  38.1     17    1247
Total:          5   19  38.2     17    1250

Percentage of the requests served within a certain time (ms)
  50%     17
  66%     19
  75%     20
  80%     21
  90%     24
  95%     26
  98%     30
  99%     32
 100%   1250 (longest request)
```